+++ CHANGELOG.md
@@ -23,6 +23,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Instance generation configuration constants.
 - Increment from 3 to 10 instance files generation for complex instances.
 - Documentation for generating instances with classic constraints only in README.
+- One Team constraint identification.
+- SUAL, WangLi, and ADA constraints detailed analysis.
 
 ### Changed
 - Updated the GUI `views.py` and `controllers.py` to fit in to our new solver solution (`ortools-fast.py` but now in a dynamic and versatile way).